Move new parts into library and update paths
