+++ src/stories/Library/material-description/material-description.scss
